requirements.txt
mongoengine
python-telegram-bot
requests
coinbase
dnspython